pyproject.toml
[project]
name = "pydantic-ai-mcp-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "eschmidt42", email = "11818904+eschmidt42@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "logfire[httpx]>=3.16.1",
    "mcp[cli]>=1.9.2",
    "openai>=1.82.1",
    "pydantic-ai-slim[mcp]>=0.2.12",
]

[project.scripts]
pydantic-ai-mcp-experiment = "pydantic_ai_mcp_experiment:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
]
